Move lovat_api.dart into lovat_api folder and refactor out GetTeams
