+++ lib/analysis_functions/match_predictor_analysis.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:scouting_dashboard_app/analysis_functions/analysis.dart';
 import 'package:http/http.dart' as http;
-import 'package:scouting_dashboard_app/reusable/lovat_api.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
 import '../constants.dart';
 
 class MatchPredictorAnalysis extends AnalysisFunction {
