+++ lib/analysis_functions/team_lookup_categories_analysis.dart
@@ -3,7 +3,7 @@ import 'package:http/http.dart' as http;
 
 import 'package:scouting_dashboard_app/analysis_functions/analysis.dart';
 import 'package:scouting_dashboard_app/constants.dart';
-import 'package:scouting_dashboard_app/reusable/lovat_api.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
 
 class TeamLookupCategoriesAnalysis extends AnalysisFunction {
   TeamLookupCategoriesAnalysis({
