+++ lib/analysis_functions/team_lookup_notes_analysis.dart
@@ -1,5 +1,5 @@
 import 'package:scouting_dashboard_app/analysis_functions/analysis.dart';
-import 'package:scouting_dashboard_app/reusable/lovat_api.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
 
 class TeamLookupNotesAnalysis extends AnalysisFunction {
   TeamLookupNotesAnalysis({
