+++ lib/reusable/models/team.dart
@@ -1,4 +1,4 @@
-import 'package:scouting_dashboard_app/reusable/lovat_api.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
 
 class Team {
   const Team({
